Implements SRP & OCP
