+++ src/main/java/id/ac/ui/cs/advprog/eshop/service/CarServiceImpl.java
@@ -1,6 +1,6 @@
 package id.ac.ui.cs.advprog.eshop.service;
 import id.ac.ui.cs.advprog.eshop.model.Car;
-import id.ac.ui.cs.advprog.eshop.repository.CarRepository;
+import id.ac.ui.cs.advprog.eshop.repository.CarRepositoryImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
@@ -10,17 +10,17 @@ import java.util.List;
 @Service
 public class CarServiceImpl implements CarService {
     @Autowired
-    private CarRepository carRepository;
+    private CarRepositoryImpl carRepositoryImpl;
 
     @Override
     public Car create(Car car){
-        carRepository.create(car);
+        carRepositoryImpl.create(car);
         return car;
     }
 
     @Override
     public List<Car> findAll() {
-        Iterator<Car> carIterator = carRepository.findAll();
+        Iterator<Car> carIterator = carRepositoryImpl.findAll();
         List<Car> allCar = new ArrayList<>();
         carIterator.forEachRemaining(allCar::add);
         return allCar;
@@ -28,18 +28,18 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public Car findById(String carId) {
-        Car car = carRepository.findById(carId);
+        Car car = carRepositoryImpl.findById(carId);
         return car;
     }
 
     @Override
     public void update(String carId, Car car) {
-        carRepository.update(carId, car);
+        carRepositoryImpl.update(carId, car);
     }
 
     @Override
     public void deleteCarById(String carId) {
-        carRepository.delete(carId);
+        carRepositoryImpl.delete(carId);
     }
 
 }
